requirements.txt
alembic==1.14.1
aniso8601==10.0.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.0
MarkupSafe==3.0.2
marshmallow==3.26.0
marshmallow-sqlalchemy==1.4.0
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.10.1
pytest==8.1.1
python-dotenv==1.0.1
pytz==2024.1
setuptools==70.3.0
six==1.16.0
SQLAlchemy==2.0.29
SQLAlchemy-serializer==1.4.1
typing_extensions==4.10.0
Werkzeug==3.1.3
wheel==0.42.0
